weather_service/requirements.txt
flask
boto3
flask-cors
python-dotenv